- Added simple statistics class - Added Polynomial.cpp and Statistics.cpp to CMakeLists.txt - Jacobian and VectorND updated to use Eigen
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@4904 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/math/Jacobian.cpp
@@ -74,8 +74,8 @@ const Jacobian rw::math::operator*(const Rotation3D<>& r, const Jacobian& jacobi
     Jacobian::Base v(jacobian.e());
     Jacobian::Base rv(v.rows(), v.cols());
 	Rotation3D<>::EigenMatrix3x3 rm = r.e();
-    for(size_t row = 0; row < v.rows()-1; row += 3 ){
-        for (size_t col = 0; col < v.cols(); col++) {
+    for(int row = 0; row < v.rows()-1; row += 3 ){
+        for (int col = 0; col < v.cols(); col++) {
 			rv.block<3,1>(row, col) = rm*v.block<3,1>(row, col);
         }
     }
